// context/types.ts
export interface Colors {
  primary: string;
  background: string;
  card: string;
  text: string;
  border: string;
  notification: string;
  surface: string;
  error: string;
  subtext: string;
  accent: string;
  buttonText: string;
  placeholder: string;
  success: string;
  warning: string;
  info: string;
}

export interface Font {
  heading: number;
  body: number;
  caption: number;
  title: number;
}

export interface Spacing {
  sm: number;
  md: number;
  lg: number;
  xl: number;
}

export interface Radius {
  sm: number;
  md: number;
  lg: number;
  xl: number;
}